@1383 Pinning down where we can forward an update to ourself.
Took 39 minutes

Took 6 minutes

+++ lucene/core/src/java/org/apache/lucene/util/IOUtils.java
@@ -195,6 +195,7 @@ public final class IOUtils {
    */
   public static void deleteFilesIgnoringExceptions(Directory dir, Collection<String> files) {
     for(String name : files) {
+      if (name == null) continue;
       try {
         dir.deleteFile(name);
       } catch (Throwable ignored) {
